+++ src/main/java/racingcar/controller/ExceptionAdvice.java
@@ -1,0 +1,39 @@
+package racingcar.controller;
+
+import java.util.Map;
+import java.util.Objects;
+import java.util.stream.Collectors;
+import org.springframework.context.support.DefaultMessageSourceResolvable;
+import org.springframework.http.HttpStatus;
+import org.springframework.validation.FieldError;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+public class ExceptionAdvice {
+
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    public Map<String, String> handle(MethodArgumentNotValidException exception) {
+        return exception.getBindingResult()
+                .getAllErrors().stream()
+                .filter(error -> Objects.nonNull(error.getDefaultMessage()))
+                .collect(Collectors.toMap(
+                        error -> ((FieldError) error).getField(), DefaultMessageSourceResolvable::getDefaultMessage)
+                );
+    }
+
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    @ExceptionHandler(IllegalArgumentException.class)
+    public String handle(IllegalArgumentException exception) {
+        return exception.getMessage();
+    }
+
+    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+    @ExceptionHandler(Exception.class)
+    public String handle(Exception exception) {
+        return exception.getMessage();
+    }
+}
